Move OhmsIndexDisplay to OralHistory::OhmsIndexComponent
