+++ spec/components/oral_history/ohms_index_component_spec.rb
@@ -1,14 +1,13 @@
 require 'rails_helper'
 
-describe OhmsIndexDisplay, type: :presenter do
+describe OralHistory::OhmsIndexComponent, type: :component do
   let(:work) { create(:oral_history_work)}
   let(:ohms_xml_path) { Rails.root + "spec/test_support/ohms_xml/duarte_OH0344.xml"}
   let(:ohms_xml) { OralHistoryContent::OhmsXml.new(File.read(ohms_xml_path))}
-  let(:ohms_index_display) { OhmsIndexDisplay.new(ohms_xml, work: work) }
+  let(:ohms_index_display) { OralHistory::OhmsIndexComponent.new(ohms_xml, work: work) }
+  let(:parsed) { render_inline ohms_index_display}
 
   it "smoke test" do
-    parsed = Nokogiri::HTML.fragment(ohms_index_display.display)
-
     expect(parsed.css("div.ohms-index-container > .ohms-index-point").count).to eq(ohms_xml.index_points.count)
   end
 
@@ -16,7 +15,6 @@ describe OhmsIndexDisplay, type: :presenter do
     let(:ohms_xml_path) { Rails.root + "spec/test_support/ohms_xml/index_hyperlinks_example.xml"}
 
     it "renders" do
-      parsed = Nokogiri::HTML.fragment(ohms_index_display.display)
       expect(parsed.css(".ohms-index-point").count).to eq(2)
 
       with_hyperlinks = parsed.at_css(".ohms-index-point[2]")
